requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.5.0,<3.0.0
openai==1.51.0
python-dotenv==1.0.0
httpx==0.24.1

# LangChain implementation dependencies
# Latest stable versions for better compatibility
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# LlamaIndex implementation dependencies  
# Latest stable versions with improved Pydantic v2 support
llama-index>=0.11.0
llama-index-llms-openai>=0.2.0
llama-index-embeddings-openai>=0.2.0